requirements.txt
matplotlib==3.7.1
mlflow==2.4.0
opencv-python==4.7.0.72
pandas==2.0.2
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
tqdm==4.65.0
transformers==4.29.2
torchlibrosa==0.1.0
h5py==3.9.0
torchsummary==1.5.1
moviepy==2.2.1
